drm: zynq_drm: crtc: Add to set encoder dpms
This patch adds to set encoder dpms in mode_set_base() in crtc.
If a monitor gets suspended, and then mode_set_base() is called,
a monitor is still off because mode_set_base() simply updates
the address of framebuffer. So search for an encoder/set dpms of
an encoder to ON is added. + other minor changes.

Signed-off-by: hyun woo kwon <hyunk@xilinx.com>