v4l: xilinx: sobel: Add interrupt support
The filter core can generate an interrupt when it has finished
processing the frame. While this is not required from a system point of
view during normal operation, it adds a trace point that can help
debugging pipeline problems. Register an interrupt handler that will
just acknowledge the interrupt and print a debug message.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>