drm: zynq_drm: drv: Remove calling drm_fbdev_cma_hotplug_event()
Remove calling drm_fbdev_dma_hotplug_event(). It was called to
pick up the right resolution from a monitor through adv7511.
adv7511 doesn't accept the first hpd signal and considers it
as a transition from off to on. This causes the initial config
from KMS uses the default resolution instead of a monitor's default
resolution.

Signed-off-by: hyun woo kwon <hyunk@xilinx.com>