media: xilinx: tpg: Configure bayer phases based on passthrough mode
For TPG v5.0, the bayer phase needs to be configured only in non passthrough
mode, otherwise the external input will be subsampled. The TPG format
is decided in dts and fixed at runtime, so that given format is used to
program the bayer phase register when passthrough mode is configured.
The bayer phase is off in passthrough mode.

Signed-off-by: Hyun Kwon <hyunk@xilinx.com>
---
v2
- This patch is added
---