[hack] media: xilinx: dma: Workaround for bytesperline
Workaround for bytesperline bug from mplayer. The version of mplayer
that was used for demo uses the bytes per line value from the driver.
Since mplayer doesn't change this value, mplayer uses the wrong bytes
per line value, for example, the default value (1920 * 4) for differnt
resolution (640x480).

The best way is to send a patch to mplayer the mailing list, and this
doesn't intend to be included in the kernel tree.

Signed-off-by: Hyun Kwon <hyunk@xilinx.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>