media: xilinx: tpg: Allocate active formats for each pad
When TPG has an input pad, the 2 try formats are stored in the file
handle. So 2 active formats are allocated in struct xtpg_device, and
this will make format handling more consistent, and help to store active
formats in struct xvip_device later as one format per pad will be needed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>